Move ComponentId create logic ang ComponentType to api package
